Role check in routes

+++ routes/web.php
@@ -20,17 +20,17 @@ use App\Http\Controllers\BidController;
 //veiling routes
 Route::get('/', [AuctionController::class, 'index'])->name('homepage');
 Route::get('/auction/list', [AuctionController::class, 'list'])->name('auction.list');
-Route::get('/auction/create', [AuctionController::class, 'create'])->name('auction.create');
-Route::post('auction/store', [AuctionController::class, 'store'])->name('auction.store');
+Route::get('/auction/create', [AuctionController::class, 'create'])->name('auction.create')->middleware('checkRole:admin');
+Route::post('auction/store', [AuctionController::class, 'store'])->name('auction.store')->middleware('checkRole:admin');
 Route::get('/auction/show/{id?}', [AuctionController::class, 'show'])->name('auction.show');
 
 // veiling\items routes
-Route::get('/auction/item/create/{id}', [ItemController::class, 'create'])->name('item.create');
-Route::post('/auction/item/store', [ItemController::class, 'store'])->name('item.store');
+Route::get('/auction/item/create/{id}', [ItemController::class, 'create'])->name('item.create')->middleware('checkRole:admin');
+Route::post('/auction/item/store', [ItemController::class, 'store'])->name('item.store')->middleware('checkRole:admin');
 Route::get('/auction/item/show/{id}', [ItemController::class, 'show'])->name('item.show');
-Route::get('/auction/item/edit/{id}', [ItemController::class, 'edit'])->name('item.edit');
-Route::post('/auction/item/update', [ItemController::class, 'update'])->name('item.update');
-Route::delete('/auction/item/destroy/{item}', [ItemController::class, 'destroy'])->name('item.destroy');
+Route::get('/auction/item/edit/{id}', [ItemController::class, 'edit'])->name('item.edit')->middleware('checkRole:admin');
+Route::post('/auction/item/update', [ItemController::class, 'update'])->name('item.update')->middleware('checkRole:admin');
+Route::delete('/auction/item/destroy/{item}', [ItemController::class, 'destroy'])->name('item.destroy')->middleware('checkRole:admin');
 
 // veiling\items\bid routes
 Route::post('/auction/item/bid/store', [BidController::class, 'store'])->name('bid.store');
